+++ truffle/src/com.oracle.truffle.api.instrumentation/src/com/oracle/truffle/api/instrumentation/InstrumentAccessor.java
@@ -40,28 +40,38 @@
  */
 package com.oracle.truffle.api.instrumentation;
 
+import java.io.InputStream;
+import java.util.Set;
+import java.util.function.Supplier;
+
+import org.graalvm.options.OptionDescriptor;
+import org.graalvm.options.OptionDescriptors;
+import org.graalvm.options.OptionValues;
+import org.graalvm.polyglot.io.MessageTransport;
+
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.Scope;
 import com.oracle.truffle.api.TruffleContext;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.impl.Accessor;
 import com.oracle.truffle.api.impl.DispatchOutputStream;
+import com.oracle.truffle.api.instrumentation.InstrumentationHandler.InstrumentClientInstrumenter;
+import com.oracle.truffle.api.instrumentation.TruffleInstrument.ContextLocalFactory;
+import com.oracle.truffle.api.instrumentation.TruffleInstrument.ContextThreadLocalFactory;
 import com.oracle.truffle.api.nodes.LanguageInfo;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
-import com.oracle.truffle.api.instrumentation.InstrumentationHandler.InstrumentClientInstrumenter;
-import org.graalvm.options.OptionDescriptor;
-import org.graalvm.options.OptionDescriptors;
-import org.graalvm.options.OptionValues;
-import org.graalvm.polyglot.io.MessageTransport;
-
-import java.io.InputStream;
-import java.util.Set;
-import java.util.function.Supplier;
 
 final class InstrumentAccessor extends Accessor {
 
     static final InstrumentAccessor ACCESSOR = new InstrumentAccessor();
+
+    static final NodeSupport NODES = ACCESSOR.nodeSupport();
+    static final SourceSupport SOURCE = ACCESSOR.sourceSupport();
+    static final JDKSupport JDK = ACCESSOR.jdkSupport();
+    static final LanguageSupport LANGUAGE = ACCESSOR.languageSupport();
+    static final EngineSupport ENGINE = ACCESSOR.engineSupport();
+    static final InteropSupport INTEROP = ACCESSOR.interopSupport();
 
     private InstrumentAccessor() {
     }
@@ -278,6 +288,24 @@ final class InstrumentAccessor extends Accessor {
                 throw InstrumentAccessor.engineAccess().invalidSharingError(currentPolyglotEngine);
             }
             return true;
+        }
+
+        @Override
+        public Object invokeContextLocalFactory(Object factory, TruffleContext truffleContext) {
+            Object result = ((ContextLocalFactory<?>) factory).create(truffleContext);
+            if (result == null) {
+                throw new IllegalStateException(String.format("%s.create is not allowed to return null.", ContextLocalFactory.class.getSimpleName()));
+            }
+            return result;
+        }
+
+        @Override
+        public Object invokeContextThreadLocalFactory(Object factory, TruffleContext truffleContext, Thread t) {
+            Object result = ((ContextThreadLocalFactory<?>) factory).create(truffleContext, t);
+            if (result == null) {
+                throw new IllegalStateException(String.format("%s.create is not allowed to return null.", ContextThreadLocalFactory.class.getSimpleName()));
+            }
+            return result;
         }
 
     }