+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotEngineImpl.java
@@ -113,6 +113,9 @@ import com.oracle.truffle.api.nodes.LanguageInfo;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.polyglot.PolyglotContextImpl.ContextWeakReference;
 import com.oracle.truffle.polyglot.PolyglotLimits.EngineLimits;
+import com.oracle.truffle.polyglot.PolyglotLocals.AbstractContextLocal;
+import com.oracle.truffle.polyglot.PolyglotLocals.AbstractContextThreadLocal;
+import com.oracle.truffle.polyglot.PolyglotLocals.LocalLocation;
 
 final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl implements com.oracle.truffle.polyglot.PolyglotImpl.VMObject {
 
@@ -136,6 +139,7 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
     private static final Map<PolyglotEngineImpl, Void> ENGINES = Collections.synchronizedMap(new WeakHashMap<>());
     private static volatile boolean shutdownHookInitialized = false;
     private static final boolean DEBUG_MISSING_CLOSE = Boolean.getBoolean("polyglotimpl.DebugMissingClose");
+    static final LocalLocation[] EMPTY_LOCATIONS = new LocalLocation[0];
 
     Engine creatorApi; // effectively final
     Engine currentApi;
@@ -194,6 +198,9 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
     final SpecializationStatistics specializationStatistics;
     final Function<String, TruffleLogger> engineLoggerSupplier;
     private volatile TruffleLogger engineLogger;
+
+    @CompilationFinal volatile StableLocalLocations contextLocalLocations = new StableLocalLocations(EMPTY_LOCATIONS);
+    @CompilationFinal volatile StableLocalLocations contextThreadLocalLocations = new StableLocalLocations(EMPTY_LOCATIONS);
 
     PolyglotEngineImpl(PolyglotImpl impl, DispatchOutputStream out, DispatchOutputStream err, InputStream in, Map<String, String> options,
                     boolean allowExperimentalOptions, boolean useSystemProperties, ClassLoader contextClassLoader, boolean boundEngine,
@@ -850,6 +857,7 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
             }
             l.initialize(context.config.limits, context);
         }
+
         if (context.config.hostClassLoader != null) {
             context.engine.customHostClassLoader.invalidate();
         }
@@ -1532,6 +1540,12 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
                 replayEvents = true;
             }
 
+            if (!replayEvents) { // is new context
+                synchronized (context) {
+                    context.initializeContextLocals();
+                }
+            }
+
             if (replayEvents && EngineAccessor.INSTRUMENT.hasContextBindings(this)) {
                 // replace events for preinitialized contexts
                 // events must be replayed without engine lock.
@@ -1708,9 +1722,9 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
 
     PolyglotThreadInfo getCachedThreadInfo(PolyglotContextImpl context) {
         if (singleThreadPerContext.isValid() && singleContext.isValid()) {
-            return context.getCachedThreadInfo(true);
+            return context.constantCurrentThreadInfo;
         } else {
-            return context.getCachedThreadInfo(false);
+            return context.currentThreadInfo;
         }
     }
 
@@ -1723,4 +1737,88 @@ final class PolyglotEngineImpl extends AbstractPolyglotImpl.AbstractEngineImpl i
             this.logLevels = new HashMap<>();
         }
     }
+
+    LocalLocation[] addContextLocals(List<? extends AbstractContextLocal<?>> newLocals, boolean invokeFactory) {
+        List<PolyglotContextImpl> aliveContexts;
+        LocalLocation[] newLocations;
+        StableLocalLocations newStableLocations;
+        synchronized (this) {
+            StableLocalLocations stableLocations = this.contextLocalLocations;
+            int index = stableLocations.locations.length;
+            LocalLocation[] locationsCopy = Arrays.copyOf(stableLocations.locations, stableLocations.locations.length + newLocals.size());
+            for (AbstractContextLocal<?> newLocal : newLocals) {
+                locationsCopy[index] = newLocal.createLocation(index);
+                newLocal.initializeLocation(locationsCopy[index]);
+                index++;
+            }
+            /*
+             * We pick up the alive contexts before we set the new context locals. So all new
+             * contexts from now on will already use the initialized locals.
+             */
+            aliveContexts = collectAliveContexts();
+            this.contextLocalLocations = newStableLocations = new StableLocalLocations(locationsCopy);
+            stableLocations.assumption.invalidate("Context local added");
+            newLocations = Arrays.copyOfRange(locationsCopy, stableLocations.locations.length, index);
+            stableLocations = this.contextLocalLocations;
+        }
+        for (PolyglotContextImpl context : aliveContexts) {
+            synchronized (context) {
+                if (context.closed || context.invalid) {
+                    continue;
+                }
+                context.resizeContextLocals(newStableLocations);
+                if (invokeFactory) {
+                    context.invokeContextLocalsFactory(context.contextLocals, newLocations);
+                }
+            }
+        }
+        return newLocations;
+    }
+
+    LocalLocation[] addContextThreadLocals(List<? extends AbstractContextThreadLocal<?>> newLocals, boolean invokeFactory) {
+        List<PolyglotContextImpl> aliveContexts;
+        LocalLocation[] newLocations;
+        StableLocalLocations newStableLocations;
+        synchronized (this) {
+            StableLocalLocations stableLocations = this.contextThreadLocalLocations;
+            int index = stableLocations.locations.length;
+            LocalLocation[] locationsCopy = Arrays.copyOf(stableLocations.locations, stableLocations.locations.length + newLocals.size());
+            for (AbstractContextThreadLocal<?> newLocal : newLocals) {
+                locationsCopy[index] = newLocal.createLocation(index);
+                newLocal.initializeLocation(locationsCopy[index]);
+                index++;
+            }
+            /*
+             * We pick up the alive contexts before we set the new context locals. So all new
+             * contexts from now on will already use the initialized locals.
+             */
+            aliveContexts = collectAliveContexts();
+            this.contextThreadLocalLocations = newStableLocations = new StableLocalLocations(locationsCopy);
+            stableLocations.assumption.invalidate("Context thread local added");
+            newLocations = Arrays.copyOfRange(locationsCopy, stableLocations.locations.length, index);
+        }
+        for (PolyglotContextImpl context : aliveContexts) {
+            synchronized (context) {
+                if (context.closed || context.invalid) {
+                    continue;
+                }
+                context.resizeContextThreadLocals(newStableLocations);
+                if (invokeFactory) {
+                    context.invokeContextThreadLocalFactory(newLocations);
+                }
+            }
+        }
+        return newLocations;
+    }
+
+    static final class StableLocalLocations {
+
+        @CompilationFinal(dimensions = 1) final LocalLocation[] locations;
+        final Assumption assumption = Truffle.getRuntime().createAssumption();
+
+        StableLocalLocations(LocalLocation[] locations) {
+            this.locations = locations;
+        }
+    }
+
 }